REFACTOR : memberId 의 타입을 Integer 에서 Long 으로 변경합니다.
일반적으로, 엔티티의 PK는 Long과 Integer 중 하나를 사용하나, 다른 엔티티와의 일관성을 유지하기 위해 Long 으로 수정합니다. 메소드의 아규먼트의 경우에는 null 에러를 피하기 위해 long 타입을 사용합니다.

+++ src/main/java/dev/pjc1991/commerce/member/point/controller/MemberPointController.java
@@ -24,7 +24,7 @@ public class MemberPointController {
     @GetMapping("/member/{memberId}/point/total")
     @ResponseBody
     @ResponseStatus(HttpStatus.OK)
-    public MemberPointTotalResponse getMemberPointTotal(@PathVariable int memberId) {
+    public MemberPointTotalResponse getMemberPointTotal(@PathVariable long memberId) {
         return memberPointService.getMemberPointTotalResponse(memberId);
     }
 
@@ -39,7 +39,7 @@ public class MemberPointController {
     @GetMapping("/member/{memberId}/point")
     @ResponseBody
     @ResponseStatus(HttpStatus.OK)
-    public Page<MemberPointEventResponse> getMemberPointEvents(@PathVariable int memberId, @RequestBody(required = false) MemberPointEventSearch search) {
+    public Page<MemberPointEventResponse> getMemberPointEvents(@PathVariable long memberId, @RequestBody(required = false) MemberPointEventSearch search) {
 
         if (search == null) {
             search = new MemberPointEventSearch();
@@ -59,7 +59,7 @@ public class MemberPointController {
     @PostMapping("/member/{memberId}/point/earn")
     @ResponseBody
     @ResponseStatus(HttpStatus.CREATED)
-    public MemberPointEventResponse earnMemberPoint(@PathVariable int memberId, @RequestBody MemberPointCreateRequest memberPointCreate) {
+    public MemberPointEventResponse earnMemberPoint(@PathVariable long memberId, @RequestBody MemberPointCreateRequest memberPointCreate) {
         memberPointCreate.setMemberId(memberId);
         return memberPointService.earnMemberPointResponse(memberPointCreate);
     }
@@ -75,7 +75,7 @@ public class MemberPointController {
     @PostMapping("/member/{memberId}/point/use")
     @ResponseBody
     @ResponseStatus(HttpStatus.CREATED)
-    public MemberPointEventResponse useMemberPoint(@PathVariable int memberId, @RequestBody MemberPointUseRequest memberPointUseRequest) {
+    public MemberPointEventResponse useMemberPoint(@PathVariable long memberId, @RequestBody MemberPointUseRequest memberPointUseRequest) {
         memberPointUseRequest.setMemberId(memberId);
         return memberPointService.useMemberPointResponse(memberPointUseRequest);
     }
